(PLATFORM-3159) Fix forum board creation
